[Master/Slave] Improvement in Master/Slave     - Improvement in Master/Slave communication protocol     - Added error handling and fault tolerance mechanisms     - Updated main.c files for both Master and Slave to incorporate new protocol changes     - Fix bugs related to communication and data processing in both Master and Slave implementations

+++ common/inc/can_protocol.h
@@ -37,7 +37,7 @@ typedef struct __attribute__((packed)) {
 
 /* Structure for RTOS CAN messages */
 typedef struct {
-    uint8_t sender_id;
+    uint32_t sender_id;
     can_frame_payload_t frame;
 } rtos_can_msg_t;
